+++ packages/atomic/src/utils/xss-utils.ts
@@ -1,17 +1,3 @@
-export function sanitize(string: string) {
-  const map: Record<string, string> = {
-    '&': '&amp;',
-    '<': '&lt;',
-    '>': '&gt;',
-    '"': '&quot;',
-    "'": '&#x27;',
-    '/': '&#x2F;',
-    '`': '&#x60;',
-  };
-  const reg = /[&<>"'/]/gi;
-  return string.replace(reg, (match) => map[match]);
-}
-
 export function filterProtocol(uri: string) {
   const isAbsolute = /^(https?|ftp|file|mailto|tel):/i.test(uri);
   const isRelative = /^\//.test(uri);